Confirm list deletions, fix duplicate click events
Modified:
 - Confirmation prompt for list delete

Fixed:
 - Duplicate click events being fired

+++ components/list.js
@@ -67,7 +67,9 @@ List.prototype.tap = function(itemIndex) {
 			break;
 
 		case "delete":
-			this.deleteEventHandler(itemIndex);
+			if (window.confirm("Delete this item?")) {
+				this.deleteEventHandler(itemIndex);
+			}
 			break;
 	}
 }
